+++ app/store/rootSaga.ts
@@ -1,6 +1,6 @@
 import { all } from "redux-saga/effects";
 
-import { messagesSaga } from "../features/messages/messagesSaga";
+import { messagesSaga } from "../features/messages/messagesSagas";
 
 export default function* rootSaga() {
   yield all([messagesSaga()]);